+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Documentation for contributing/develop
 - Documentation for JS classes/methods
+- Unit test framework Mocha
+
+### Changed
+- Merged webpack configuration files into a single one
 
 
 ## 0.1.0 - 2017-12-20